fix: test case for contribution and rename folder
