+++ src/bondContribution/spectraPreparation.js
